walkabout: No heap allocations for top-level Walk function.
The previous approach to implementing the top-level WalkInterface() function
would cause the interface value to escape to the heap. This new code-gen
approach grabs the interface's value pointer and passes that into
Engine.Execute(), instead of creating an KindInterface frame.

The benchmark and no-malloc test are updated to ensure that the top-level
WalkInterface() function is called.