walkabout: Use new golang.org/x/tools/packages library
The packages library does precisely what generation does, but in a fashion
guaranteed to be portable across module and build systems.

In order to facilitate testing, the re-parse step is no longer built into
generation, and the abstract accessor API is tweaked to allow multiple
generated APIs to co-exist in the same package.