walkabout: Add linting and various build cleanups.
This change adds golint and staticcheck to the Makefile. Per the documentation
for go modules, a `tools.go` has been added with a no-build build directive and
dummy imports on the tools.

The demo `go:generate` directives are tweaked to use `go run`, instead of
requiring a `go install` to be performed.